# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API",
        "description": "This is the API documentation for my Invoices application.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/user/list": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "List Users",
                "description": "Returns a list of users based on filter parameters. Requires authentication.",
                "operationId": "30445baac79ff5f15e8e33563aad2af9",
                "requestBody": {
                    "description": "Filtering, sorting, and pagination parameters",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaginatedListFilter"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get User",
                "description": "Returns a single user by ID. Requires authentication.",
                "operationId": "d8f696420bea235f4916124c9ba1c507",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update User",
                "description": "Updates an existing user. Requires authentication and proper authorization.",
                "operationId": "5e09c26a0abb2b1bf040badb2d9c69ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User data for update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete User",
                "description": "Deletes a user by ID. Requires admin privileges.",
                "operationId": "f8d7aea516ee464d0afee90f3a1a1722",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create User",
                "description": "Creates a new user. Requires admin privileges.",
                "operationId": "089999749fbdec98922de2f74ba8d540",
                "requestBody": {
                    "description": "User data to create a new user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Filter": {
                "title": "Filter",
                "description": "A single filter condition.",
                "required": [
                    "field",
                    "operator"
                ],
                "properties": {
                    "field": {
                        "description": "Field name to filter on.",
                        "type": "string",
                        "example": "user_id"
                    },
                    "value": {
                        "description": "Value to filter by.",
                        "type": "string",
                        "example": "36"
                    },
                    "fieldType": {
                        "description": "Type of the field.",
                        "type": "string",
                        "enum": [
                            "text",
                            "date",
                            "number",
                            "boolean",
                            "set",
                            "array"
                        ],
                        "example": "number"
                    },
                    "operator": {
                        "description": "Operator for the filter condition.",
                        "type": "string",
                        "enum": [
                            "contains",
                            "notContains",
                            "equals",
                            "notEqual",
                            "startsWith",
                            "endsWith",
                            "blank",
                            "notBlank",
                            "greaterThan",
                            "greaterThanOrEqual",
                            "lessThan",
                            "lessThanOrEqual",
                            "inRange"
                        ],
                        "example": "equals"
                    },
                    "rangeValue": {
                        "description": "Secondary value for range filters (if applicable).",
                        "type": "string",
                        "example": "50",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginatedListFilter": {
                "title": "Paginated List Filter",
                "description": "Parameters for filtering, sorting, and pagination.",
                "required": [
                    "filters",
                    "sort",
                    "limit"
                ],
                "properties": {
                    "filters": {
                        "description": "Array of filter objects",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Filter"
                        }
                    },
                    "sort": {
                        "$ref": "#/components/schemas/Sort"
                    },
                    "limit": {
                        "description": "Maximum number of records to return",
                        "type": "integer",
                        "example": 15
                    }
                },
                "type": "object"
            },
            "Sort": {
                "title": "Sort",
                "description": "Sorting options for a query.",
                "required": [
                    "field",
                    "direction"
                ],
                "properties": {
                    "field": {
                        "description": "Field to sort by.",
                        "type": "string",
                        "example": "invoice_number"
                    },
                    "direction": {
                        "description": "Sort direction.",
                        "type": "string",
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "example": "asc"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User Model",
                "description": "User entity for authentication and user management",
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier for the user",
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the user",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Email address of the user",
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "role": {
                        "description": "Role assigned to the user",
                        "type": "string",
                        "enum": [
                            "user",
                            "admin"
                        ],
                        "example": "user"
                    },
                    "password": {
                        "description": "User's password (write-only)",
                        "type": "string",
                        "writeOnly": true
                    },
                    "email_verified_at": {
                        "description": "Timestamp when the user's email was verified",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-15T13:45:30Z",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Timestamp when the user was created",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2023-01-15T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Timestamp when the user was last updated",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2023-02-01T08:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}